+++ src/utils/Routes/Prefix.tsx
@@ -1,0 +1,4 @@
+// Desc: Prefix for the routes
+const prefix = '/react-wheels';
+
+export { prefix };
