+++ src/utils/Routes/index.tsx
@@ -1,0 +1,4 @@
+import { searchRoute, routesList } from './Routes';
+import { prefix } from './Prefix';
+
+export { searchRoute, routesList, prefix };
